Moved fixture placed_order_result to conftest.py
