+++ tests/core/conftest.py
@@ -57,6 +57,57 @@ def acccount_result() -> dict:
 
 
 @pytest.fixture
+def placed_order_result():
+    """fixture with data of an placed order
+    """
+    return {
+        "time": "2022-04-02T18:10:54.688+00:00",
+        "mode": "paper",
+        "status": "ok",
+        "results": {
+            "created_at": "2022-04-02T18:10:54.613+00:00",
+            "id": "ord_abcdefghijklmnopqrstuvwxyz12345678",
+            "status": "inactive",
+            "regulatory_information": {
+                "costs_entry": 0,
+                "costs_entry_pct": "0.00%",
+                "costs_running": 0,
+                "costs_running_pct": "0.00%",
+                "costs_product": 0,
+                "costs_product_pct": "0.00%",
+                "costs_exit": 0,
+                "costs_exit_pct": "0.00%",
+                "yield_reduction_year": 0,
+                "yield_reduction_year_pct": "0.00%",
+                "yield_reduction_year_following": 0,
+                "yield_reduction_year_following_pct": "0.00%",
+                "yield_reduction_year_exit": 0,
+                "yield_reduction_year_exit_pct": "0.00%",
+                "estimated_holding_duration_years": "5",
+                "estimated_yield_reduction_total": 0,
+                "estimated_yield_reduction_total_pct": "0.00%",
+                "KIID": "text",
+                "legal_disclaimer": "text"
+            },
+            "isin": "US02079K3059",
+            "expires_at": "2022-04-04T21:59:00.000+00:00",
+            "side": "buy",
+            "quantity": 1,
+            "stop_price": None,
+            "limit_price": None,
+            "venue": "xmun",
+            "estimated_price": 25395000,
+            "estimated_price_total": 25395000,
+            "notes": None,
+            "charge": 0,
+            "chargeable_at": None,
+            "key_creation_id": "apk_keykeykeykeykeykeykeykeykeykeykeyk",
+            "idempotency": None
+        }
+    }
+
+
+@pytest.fixture
 def account(mocker) -> Account:
     def mock_fetch_state(self):
         return
